move tests on operators in folder.
