+++ tests/operators/test_batch.py
