+++ tests/operators/test_stacked.py
