+++ tests/operators/test_stacked_gpu.py
